docs/codes/requirements.txt
torch>=2.1.0
transformers==4.39.3
datasets==2.18.0
evaluate==0.4.1
trl==0.8.6
accelerate==0.27.2
peft==0.10.0
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0
pyarrow>=14.0.0
sentence-transformers>=2.6.0
huggingface-hub>=0.20.0
openai-whisper>=20231117
gradio>=4.0.0
soundfile>=0.12.1
pyspellchecker>=0.7.2
TTS>=0.22.0
sentencepiece>=0.1.99
protobuf>=4.25.0
tqdm>=4.66.0
happytransformer>=3.0.0
gramformer>=1.0.0
